chore: use pnpm and workspaces; fix: workspaces support
## Summary
Initial setup was fine when there were 2-3 packages in the repo. However, over time more packages were added, and also I am planning to add react-native-core to this repo as a submodule in the near future. react-native-core, react-native-harmony, hvigor-plugin, and cli need to be deployed together so that the version is the same for them in order to let users know which packages are compatible with each other.

The current setup has also many flaws, e.g., the tester's verify scripts verify the entire repository but they should verify only itself, packages need to be uninstalled, packed and installed after any change, react-native-core must be deployed every time it is changed because it isn't installed locally. This MR solves those issues.

Initially, I tried using NPM workspaces, however, the node_modules structure seems to be incompatible with react-native. Also, NPM doesn't provide an option to change that. Yarn node_modules structure should be compatible, however, I had issues with yarn (3.6.4) when installing packages again after the first installation (the resolution step was super slow). I decided to try pnpm, and it feels fast and reliable. I've managed to set up monorepo with it. In this repo, pnpm is configured to use the yarn node_modules hoisting approach so potential switching to yarn should be easy.

This MR also speeds up the verify script. pnpm dev runs multiple servers concurrently. This will be more useful when the test-kit library is used here because it depends on the hdc server running in the background.

The deployment script needs to be recreated after adding react-native-core as a submodule.

## Changes
- added support for symlinks and monorepos
- added `--hermersc-dir` argument to bundle-harmony

## Test Plan
Run this project according to project-setup.md instructions.

## Checklist
- [x] ...
- [x] hoist commands
- [x] update verify
- [x] update git hooks setup
- [x] update instructions
- [x] update tester-ecosystem
- [x] require verify script
- [x] ~update deployment script~ - create issue for that, depends on adding rn core to monorepo
- [x] clean up after yarn/npm
- [x] fix bundle-harmony
- [x] check production mode
- [x] test setup on fresh repo
- [x] ~use workspace protocol?~ it will be needed after adding react native core to monorepo
- [x] add react-native-core - create issue for that
- [x] clean up
- [x] self review

## Issue Links
Closes https://gl.swmansion.com/rnoh/react-native-harmony/-/issues/1487

See merge request rnoh/react-native-harmony!1804

+++ packages/react-native-harmony-cli/src/commands/bundle-harmony.ts
@@ -6,10 +6,11 @@ import { RunBuildOptions as BuildOptions } from 'metro';
 import MetroServer from 'metro/src/Server';
 import pathUtils from 'path';
 import { getAssetDestRelativePath } from '../assetResolver';
-import { ConfigT as MetroConfig } from 'metro-config';
 import { execSync } from 'child_process';
 import { Logger } from '../io';
 import { DescriptiveError } from '../core';
+
+type MetroConfig = Awaited<ReturnType<typeof Metro.loadConfig>>;
 
 const HERMESC_BIN_DIR = {
   darwin: 'osx-bin',
@@ -69,6 +70,12 @@ export const commandBundleHarmony: Command = {
       parse: (val: string) => val !== 'false',
     },
     {
+      name: '--hermesc-dir <path>',
+      description:
+        'Path to hermesc directory. hermesc is used only when generating a HBC bundle.',
+      default: HERMESC_PATH_PREFIX,
+    },
+    {
       name: '--hermesc-options <string...>',
       description:
         'Used only when generating a HBC bundle. Additional options to pass to hermesc when generating HBC bundle. Example: --hermesc-options O g reuse-prop-cache.',
@@ -76,7 +83,6 @@ export const commandBundleHarmony: Command = {
     },
   ],
   func: async (argv, config, args: any) => {
-    // input
     const logger = new Logger();
     try {
       const bundleOutput: string =
@@ -100,16 +106,15 @@ export const commandBundleHarmony: Command = {
       };
       const metroConfig = await loadMetroConfig(args.config);
 
-      // processing
       const bundle = await createBundle(metroConfig, buildOptions);
       const assets = await retrieveAssetsData(metroConfig, buildOptions);
 
-      // output
       if (shouldGenerateHbcBundle) {
         await saveHbcBundle(
           logger,
           bundle.code,
           bundleOutput,
+          args.hermescDir,
           args.hermescOptions
         );
       } else {
@@ -149,6 +154,7 @@ async function saveHbcBundle(
   logger: Logger,
   bundleCode: string,
   bundleOutput: Path,
+  hermescPath: string,
   hermescOptions: string[] | undefined
 ) {
   if (
@@ -158,12 +164,20 @@ async function saveHbcBundle(
   ) {
     throw new Error(`Unsupported platform ${process.platform} (⊙︿⊙)`);
   }
-  const resolvedHermescPath = pathUtils.join(
-    HERMESC_PATH_PREFIX,
+  const resolvedHermescPath = pathUtils.resolve(
+    hermescPath,
     HERMESC_BIN_DIR[process.platform as OSType],
     'hermesc'
   );
-  await fse.ensureFile(resolvedHermescPath);
+
+  if (!fs.existsSync(resolvedHermescPath)) {
+    throw new DescriptiveError({
+      whatHappened: `Couldn't find hermesc at ${resolvedHermescPath}`,
+      whatCanUserDo: [
+        'Find hermesc dir and provide it by using --hermesc-dir argument. hermesc dir should be in react-native package.',
+      ],
+    });
+  }
   await fse.writeFile(TMP_BUNDLE_PATH, bundleCode);
   const hbcFilePath = bundleOutput.endsWith('.hbc')
     ? bundleOutput
@@ -285,7 +299,7 @@ function copyFiles(logger: Logger, fileDestBySrc: Record<Path, Path>) {
 function copyFile(
   src: string,
   dest: string,
-  onFinished: (error: Error) => void
+  onFinished: (error?: Error) => void
 ): void {
   const destDir = pathUtils.dirname(dest);
   fs.mkdir(destDir, { recursive: true }, (err?) => {
